+++ client/pages/Wehike.tsx
@@ -20,19 +20,6 @@ export default function WeHike() {
       &quot;We Hike&quot; is designed to address these challenges. The platform allows users to sign in, save their favorite hikes, check real-time weather for specific areas, and stay updated on trail conditions.
       </p>
       <div className='flex justify-center'>
-        {/* {darkMode ?
-      <img
-       src='/images/project-photos/wehikewidetransp.png'
-        alt='we-hike project'
-        className='max-w-[90rem] max-h-[80rem] rounded-lg mb-16'>
-      </img>
-      :
-      <img
-      src='/images/project-photos/wide.png'
-       alt='we-hike project'
-       className='max-w-[90rem] max-h-[80rem] rounded-lg mb-16'>
-     </img>
-      } */}
         <img
       src='/images/mockups/wehikelg.png'
        alt='we-hike project'
